backend/requirements.txt
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
backcall==0.2.0
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.0.1
constantly==15.1.0
cryptography==39.0.0
daphne==4.0.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.1.5
django-cors-headers==3.13.0
django-extensions==3.2.3
django-filter==22.1
django-rest-framework==0.1.0
django-shell-plus==1.1.7
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
executing==1.2.0
fancycompleter==0.9.1
geographiclib==2.0
geopy==2.3.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
install==1.3.5
ipython==8.14.0
jedi==0.18.2
matplotlib-inline==0.1.6
msgpack==1.0.5
oauthlib==3.2.2
parso==0.8.3
pdbpp==0.10.3
pexpect==4.8.0
pickleshare==0.7.5
pipreqs==0.4.11
prompt-toolkit==3.0.38
psycopg2==2.9.6
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.15.1
PyJWT==2.7.0
pyOpenSSL==23.2.0
pyrepl==0.9.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.7.1
redis==4.5.5
requests==2.28.2
requests-oauthlib==1.3.1
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
traitlets==5.9.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.6.3
urllib3==1.26.14
wcwidth==0.2.6
wmctrl==0.4
yarg==0.1.9
zope.interface==6.0
